fixup! fixup! fixup! fixup!  [IMP] joint_buying_product
