+++ joint_buying_product/readme/ROADMAP.rst
@@ -1,0 +1,2 @@
+This module doesn't handle variants features.
+All features are implemented in the ``product.product`` level.
