+++ joint_buying_product/views/view_product_template.xml
@@ -13,11 +13,9 @@ License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
         <field name="arch" type="xml">
             <xpath expr="//field[@name='sale_ok']/.." position="before">
                <div>
-                    <field name="is_joint_buying" invisible="1"/>
-                    <field name="joint_buying_partner_id"
-                        context="{'joint_buying': 1}"
-                        attrs="{'invisible': [('is_joint_buying', '=', False)], 'required': [('is_joint_buying', '=', True)]}"
-                        />
+                    <!-- TODO set invisible -->
+                    <field name="is_joint_buying"/>
+                    <label for="is_joint_buying"/>
                 </div>
             </xpath>
         </field>
